pyproject.toml
[project]
name = "connected-papers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "adapters>=1.2.0",
    "arxiv>=2.2.0",
    "faiss-cpu>=1.12.0",
    "ipython>=8.37.0",
    "nbformat>=5.10.4",
    "networkx>=3.4.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "sentence-transformers>=5.1.0",
    "tqdm>=4.67.1",
]
